+++ src/api/helpers.ts
@@ -1,0 +1,5 @@
+export function assertOk(result: { ok: boolean; message: string }): void {
+  if (!result.ok) {
+    throw new Error(result.message);
+  }
+}
