--- src/server/root-router.ts
@@ -1,8 +1,0 @@
-import { router } from './trpc';
-import { setupRouter } from './routers/setup-router';
-
-export const appRouter = router({
-  setup: setupRouter,
-});
-
-export type AppRouter = typeof appRouter;
