--- src/server/trpc.ts
@@ -1,12 +1,0 @@
-import { initTRPC } from '@trpc/server';
-
-export interface TrpcContext {}
-
-export function createTrpcContext(): TrpcContext {
-  return {};
-}
-
-const t = initTRPC.context<TrpcContext>().create();
-
-export const router = t.router;
-export const publicProcedure = t.procedure;
