+++ src/web/components/ui/input.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '#lib/utils';
 
 export function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
   return (
