+++ src/web/components/ui/label.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '#lib/utils';
 
 export function Label({ className, ...props }: React.ComponentProps<'label'>) {
   return (
