+++ src/web/components/ui/select.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from "lucide-react"
 import { Select as SelectPrimitive } from "radix-ui"
 
-import { cn } from "@/lib/utils"
+import { cn } from '#lib/utils'
 
 function Select({
   ...props
