+++ src/web/routes/api/auth/$.ts
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { auth } from '../../../../server/auth';
+import { auth } from '#server/auth';
 
 export const Route = createFileRoute('/api/auth/$')({
   server: {
